Add files via upload
Wand_tapping contains the code for a wizards book that has 4 reed switches hidden under runes on it's cover, a multi-colored LED in the center, and a solenoid keeping it shut.  By tapping a particular combo with your magic wand (with magnet embedded) it causes the LED to flicker different colors and either unlock the solenoid or a secret door across the room.  This has been installed in two escape rooms - one in Richmond, VA and the other in NY and has been very successfully enjoyed in both.